move ttl cache class to helpers dir
